third_party/blink/web_tests/external/wpt/mathml/relations/html5-tree/tabindex-001.tentative-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] default and invalid values for a without href
  assert_equals: no attribute expected 0 but got -1
[FAIL] default and invalid values for a with href
  assert_equals: no attribute expected 0 but got -1
Harness: the test ran to completion.
